Feature/vulkan (#22)
* Added Vulkan dependencies

* Vulkan initialization done

* Creating virtual/physical device objects complete

* Restructured the engine instance code

- Added surface support for window class
- Queues are getting created

* Created a bunch of new graphics classes

* Fixed Vulkan issues on Windows

* More structural changes

* Finally some on-screen action

* Added additional dependencies for shader compilation

* Initial working shader compiler and test renderer

* More work on Vulkan shaders

* Can render a triangle now

* Command lists can now be requested in a multi threaded renderer

* Can now create and fill buffers with data

* More work on Vulkan backend

* Made a number of changes which should make mesh drawing possible

* Fixed rendering issue, Sponza scene now visible in Vulkan renderer

* Working towards using descriptors in shaders

* Full parsing of shader reflection data for descriptors

* Uniform buffers can now be used in tandem with descriptors

* Sampler and images

* Textured models work now

* Mipmapping and anisotropic filtering

* Swap chain recreation on window resize

* ImGui works now

* Added multi buffer

* Fixed many of the code warnings

* New graphics API object memory management

* Working on render pass and changed initialization

* Got a working secondary render pass

* Compute shader are working now

* Barriers work and image copying works

* Organised the test renderer a bit better

* More convenience, better render passes

* Ported the profiler to Vulkan and made it multi-thread capable

* Removed OpenGL dependency

* Resources now also keep track of their access masks

* Implemented frame buffers

* Reintegrated demo app

* Removed old Material

* Cubemaps should be working now

* Brought the render target up-to-date

* Removed the old framebuffer

* Removed old texture class

* Fixed Windows issues

* Removed dependencies

* New pipeline manager, removed old texture class

* Beginning work on renderers

* OpqueRenderer is working again

* Working on new direct light renderer

* Working on porting the renderers

* Fixed shader issue

* Shadows are kind of working

* Found the shadow issue

* Shadows are working correctly now

* Changes to how the old buffer class works

* Fixed a good amount of code smells

* Fixed few bugs/hot shader reload is working now

* Refactored namespaces

* Working on the ray tracing helper

* Fixed path tracing issues

* Improved memory barriers for ray tracing helper

* Some small tweaks and changes

* Working dynamic uniform buffer bindings

* Turn off validation layer

* Working on the vertex array

* Disabling the shader hot reload improves performance

* A few smaller fixes to CMake files

* Working vertex arrays

* Skybox rendering

* Probe filtering works again

* Indirect lighting shader works again

* DDGI is working again

* Removed old shader classes

* Fixed a bunch of PR warnings

* Work on TAA

* Fixed temporal anti-aliasing

* Images can now be read back from to the CPU

* Removed duplicate code/Fixed MacOS

* Working on AO

* RTAO is working again

* Reflections work again

* Fixed a few issues

* Volumetric renderer is working again

* DDGI probe debugging available again

* Text renderer is working again

* Got the atmosphere renderer working again

* Fixed the probe generation of the atmosphere

* Explicit memory cleanup when loading new scene

* Added licenses to THIRDPARTY.md/Fixed stuff

* Volumetric clouds are working again

* Cleanup of glslang use

* Updated build/release pipeline vcpkg commitsh

* Update build.yml

* Further cleanup/reverted vcpkg committish

* Fix CMake dependency issue

* Fixed some bugs and got rid of one downscale operation

* Screen-space shadow renderer working again

* Vsync should work now

* Fixed swap chain semaphore issue

* Texture renderer is working again

* Trying to fix spirv-tools build issue by downgrading

* Ported skybox renderer to compute shader

* Update build.yml

* Fix build and release pipeline

* Fixed some PR warnings

---------

+++ src/engine/renderer/helper/GeometryHelper.h
@@ -27,7 +27,7 @@ namespace Atlas {
 				static void GenerateSphere(uint32_t rings, uint32_t segments, std::vector<uint32_t>& indices,
 					std::vector<vec3>& vertices, uint32_t* indexCount, uint32_t* vertexCount);
 
-				static int8_t rectangleVertices[];
+				static float rectangleVertices[];
 
 				static float cubeVertices[];
